+++ app/controllers/forest_liana/resources_controller.rb
@@ -27,8 +27,9 @@ module ForestLiana
       getter = ForestLiana::ResourceGetter.new(@resource, params)
       getter.perform
 
+      record = getter.record.becomes(@resource)
       render serializer: nil, json:
-        serialize_model(getter.record, include: includes(getter))
+        serialize_model(record, include: includes(getter))
     end
 
     def create
@@ -39,8 +40,9 @@ module ForestLiana
         render serializer: nil, json: JSONAPI::Serializer.serialize_errors(
           creator.errors), status: 400
       elsif creator.record.valid?
+        record = creator.record.becomes(@resource)
         render serializer: nil,
-          json: serialize_model(creator.record, include: record_includes)
+          json: serialize_model(record, include: record_includes)
       else
         render serializer: nil, json: JSONAPI::Serializer.serialize_errors(
           creator.record.errors), status: 400
@@ -55,8 +57,9 @@ module ForestLiana
         render serializer: nil, json: JSONAPI::Serializer.serialize_errors(
           updater.errors), status: 400
       elsif updater.record.valid?
+        record = updater.record.becomes(@resource)
         render serializer: nil,
-          json: serialize_model(updater.record, include: record_includes)
+          json: serialize_model(record, include: record_includes)
       else
         render serializer: nil, json: JSONAPI::Serializer.serialize_errors(
           updater.record.errors), status: 400
@@ -94,7 +97,8 @@ module ForestLiana
     end
 
     def render_jsonapi getter
-      render serializer: nil, json: serialize_models(getter.records,
+      records = getter.records.map { |record| record.becomes(@resource) }
+      render serializer: nil, json: serialize_models(records,
         include: includes(getter), count: getter.count, params: params)
     end
   end