+++ app/services/forest_liana/schema_utils.rb
@@ -26,7 +26,11 @@ module ForestLiana
         if model.abstract_class?
           model_found = self.find_model_from_abstract_class(model, collection_name)
         elsif ForestLiana.name_for(model) == collection_name
-          model_found = model.base_class
+          if self.sti_child?(model)
+            model_found = model
+          else
+            model_found = model.base_class
+          end
         end
 
         break if model_found