+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - STI Models - Support STI Models with a collection for the parent model and each child model, with automatic segments on the parent collection.
 
+### Changed
+- Collections Names - Collection names are now based on the model name.
+
 ## RELEASE 1.9.8 - 2017-11-16
 ### Added
 - Stripe Integration - Allow users to display Stripe records in the Details view.