+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+
+## RELEASE 2.0.0 - 2017-11-24
 ### Added
 - STI Models - Support STI Models with a collection for the parent model and each child model, with automatic segments on the parent collection.